fix: regular runner names
